requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
Django==5.0.2
djangorestframework==3.14.0
executor==23.2
fasteners==0.19
humanfriendly==10.0
idna==3.6
MarkupSafe==2.1.5
property-manager==3.0
pyreadline3==3.4.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
verboselogs==1.7
Werkzeug==3.0.1
